% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSex.R
\name{checkSex}
\alias{checkSex}
\title{Check sex from sequencing data}
\usage{
checkSex(files, mc.cores, quiet = FALSE)
}
\arguments{
\item{files}{A folder path where the .bam files are or a vector of file
paths. This function searches in recursive folders.}

\item{mc.cores}{Number of cores to be used with this function. If there are
more cores than samples, the number of cores will be limited to the number
of samples. By default set to 1.}

\item{quiet}{Should the function not inform about the status of the process.
By default is FALSE.}
}
\value{
A checkSex object that contains the trimmed mean coverage in chrX and
chrY in the analyzed samples, and its classification as males and females.
}
\description{
Check sex from sequencing data
}
\examples{
\dontrun{
checkSex(resMADloy)}
}
