% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLOY.R
\name{getLOY}
\alias{getLOY}
\title{Detection of mosaic Loss of chromosome Y in MADseqLOY data}
\usage{
getLOY(MADseqLOY, coef = 3, ...)
}
\arguments{
\item{coef}{this determines how far the outliers are considered.}

\item{...}{Other parameters.}

\item{x}{A MADseqLOY object.}
}
\value{
An object of class 'LOY' that summarizes the LOY events detected in
the analyzed samples
}
\description{
Detection of mosaic Loss of chromosome Y in MADseqLOY data
}
\examples{
\dontrun{
getLOY(resMADseqloy)}
}
