% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBam.R
\name{processBam}
\alias{processBam}
\title{Calculation of the target and reference coverage.}
\usage{
processBam(
  file,
  bins,
  n.bases.min,
  mappability.min,
  blacklist.max,
  correction,
  span,
  family,
  normalization,
  target,
  ref,
  trim.target,
  trim.ref
)
}
\arguments{
\item{file}{File path of the bam file to be processed.}

\item{bins}{An AnnotatedDataFrame object with the annotation of all the bins.}

\item{n.bases.min}{A numeric scalar between 0 and 100 specifying the minimum
percentage of bases characterized (wihout N).}

\item{mappability.min}{A numeric scalar between 0 (no reads filtered) and 100
(maximum reads filtered) specifying the minimum mappability to filter bins.}

\item{blacklist.max}{A numeric scalar between 0 (no overlap) and 100 (maximum
overlap) specifying the maximum overlap between bins and blacklist elements.}

\item{correction}{A character vector specifying which variables to include in
the correction. Can be c("gc", "mappability"), "gc", "mappability", or
"None".}

\item{span}{If correction object different to "None", the parameter alpha
which controls the degree of smoothing in the read counts correction.}

\item{family}{If correction object different to "None", the loess family. if
"gaussian" fitting is by least-squares, and if "symmetric" a re-descending M
estimator is used with Tukey's biweight function.}

\item{normalization}{A character string specifying the normalization method
for the read counts. Choices are "mean", "median", or "None".}

\item{target}{A GenomicRanges object with the query region of interest.}

\item{ref}{A GenomicRanges object with the reference region.}

\item{trim.target}{The fraction (0 to 0.5) of observations to be trimmed
when calculating the coverage of the target region.}

\item{trim.ref}{The fraction (0 to 0.5) of observations to be trimmed
when calculating the coverage of the reference region.}
}
\value{
A list with the mean and the standard deviation for the reference and
the target region coverages.
}
\description{
Calculation of the target and reference coverage.
}
\examples{
\dontrun{
processBam(file=file.path, bins=getBinAnnotations(100,"hg19"), n.bases.min=30,
mappability.min=50, blacklist.max=0, correction=c("gc","mappability"), 
span=0.75, family="gaussian", normalization="median", target=GRangestarget, 
ref=GRangesref, trim.target=0, trim.ref=0.25)}
}
